# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentary of the Rubik's Market Laravel project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/admin/user": {
            "get": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Get User List",
                "operationId": "3f60835ebb2934231a8e8f63e0af45a7",
                "responses": {
                    "200": {
                        "description": "List of User resources",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserListResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user": {
            "post": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Create a new user",
                "operationId": "f2aa988117521168f93651927dff05ce",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "User's full name",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "phonenumber",
                        "in": "query",
                        "description": "User's phone number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "meli_code",
                        "in": "query",
                        "description": "User's national ID code",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "image",
                        "in": "query",
                        "description": "User's profile image in URL or base64 format",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "binary"
                        }
                    },
                    {
                        "name": "cart_number",
                        "in": "query",
                        "description": "User's card number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "home_number",
                        "in": "query",
                        "description": "User's home phone number",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        }
                    },
                    {
                        "name": "is_superuser",
                        "in": "query",
                        "description": "Specifies if the user is a superuser",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User's email address",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "email",
                            "maxLength": 255
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 101
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input or missing required fields",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Bad request: Missing required fields"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserListResource": {
                "title": "User List Resource",
                "description": "Schema for paginated list of users",
                "properties": {
                    "current_page": {
                        "description": "The current page number",
                        "type": "integer",
                        "example": 1
                    },
                    "data": {
                        "description": "List of user records",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "description": "User ID",
                                    "type": "integer",
                                    "example": 1
                                },
                                "name": {
                                    "description": "Name of the user",
                                    "type": "string",
                                    "example": "Admin"
                                },
                                "email": {
                                    "description": "Email address of the user",
                                    "type": "string",
                                    "format": "email",
                                    "example": "jww@email.com"
                                },
                                "phonenumber": {
                                    "description": "Phone number of the user",
                                    "type": "integer",
                                    "example": 9209293499
                                },
                                "image": {
                                    "description": "Profile photo path",
                                    "type": "string",
                                    "example": "ProfilePhoto/CScWcYdpROEVjNMrHkHbSyJIdpOUZB2n19PVlV9Y.jpg"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "first_page_url": {
                        "description": "URL to the first page",
                        "type": "string",
                        "format": "url",
                        "example": "http://api.example.com/users?page=1"
                    },
                    "from": {
                        "description": "The starting index of records on the current page",
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "description": "The last page number",
                        "type": "integer",
                        "example": 10
                    },
                    "last_page_url": {
                        "description": "URL to the last page",
                        "type": "string",
                        "format": "url",
                        "example": "http://api.example.com/users?page=10"
                    },
                    "links": {
                        "description": "Pagination navigation links",
                        "type": "array",
                        "items": {
                            "properties": {
                                "url": {
                                    "description": "URL of the pagination link",
                                    "type": "string",
                                    "format": "url",
                                    "example": "http://api.example.com/users?page=2",
                                    "nullable": true
                                },
                                "label": {
                                    "description": "Label for the pagination link",
                                    "type": "string",
                                    "example": "Next"
                                },
                                "active": {
                                    "description": "Whether the link is active or not",
                                    "type": "boolean",
                                    "example": true
                                }
                            },
                            "type": "object"
                        }
                    },
                    "next_page_url": {
                        "description": "URL to the next page, if available",
                        "type": "string",
                        "format": "url",
                        "example": "http://api.example.com/users?page=2",
                        "nullable": true
                    },
                    "path": {
                        "description": "Base path for the API endpoint",
                        "type": "string",
                        "format": "url",
                        "example": "http://api.example.com/users"
                    },
                    "per_page": {
                        "description": "Number of records per page",
                        "type": "integer",
                        "example": 15
                    },
                    "prev_page_url": {
                        "description": "URL to the previous page, if available",
                        "type": "string",
                        "format": "url",
                        "example": null,
                        "nullable": true
                    },
                    "to": {
                        "description": "The ending index of records on the current page",
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "description": "Total number of records",
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}